Refactor bUnit Tests
