+++ sample-tests/Pizzeria.Store.Components.Tests.Unit/Pizzeria.Store.Components.Tests.Unit.csproj
@@ -6,7 +6,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\sample\Pizzeria.Store.Web\Pizzeria.Store.Web.csproj" />
+    <ProjectReference Include="..\..\sample\Pizzeria.Store.Components\Pizzeria.Store.Components.csproj" />
   </ItemGroup>
 
 </Project>
