+++ sample-tests/Pizzeria.Store.Web.Tests.Unit/MainLayoutTests.cs
@@ -1,16 +1,34 @@
 using Bunit;
+using Correlate;
 using FluentAssertions;
+using Fluxor;
+using MediatR;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using Moq;
+using MudBlazor.Services;
+using Pizzeria.Store.Components;
+using Pizzeria.Store.StateManagement.Orders;
+using Xunit;
 
 namespace Pizzeria.Store.Web.Tests.Unit;
 
 public class MainLayoutTests : TestContext
 {
+    public MainLayoutTests()
+    {
+        this.Services.AddSingleton(Mock.Of<IMediator>());
+        this.Services.AddSingleton(Mock.Of<ICorrelationContextAccessor>());
+        this.Services.AddSingleton(Mock.Of<ILogger<OrdersEffects>>());
+        this.Services.AddMudServices();
+        this.Services.AddFluxor(o => o.ScanAssemblies(typeof(OrdersState).Assembly));
+
+        this.JSInterop.Mode = JSRuntimeMode.Loose;
+    }
+
     [Fact]
     public void MainLayout_WhenRendered_ShowsAppBarWithTitle()
     {
-        // Arrange
-        this.Setup();
-
         // Act
         var component = this.RenderComponent<MainLayout>();
 
@@ -21,9 +39,6 @@ public class MainLayoutTests : TestContext
     [Fact]
     public void MainLayout_WhenRendered_ShowsSignOutButton()
     {
-        // Arrange
-        this.Setup();
-
         // Act
         var component = this.RenderComponent<MainLayout>();
 
@@ -34,9 +49,6 @@ public class MainLayoutTests : TestContext
     [Fact]
     public void MainLayout_WhenRendered_ShowsSignOutFormWithCorrectAction()
     {
-        // Arrange
-        this.Setup();
-
         // Act
         var component = this.RenderComponent<MainLayout>();
