--- sample-tests/Pizzeria.Store.Web.Tests.Unit/OrderPageTests.cs
@@ -1,90 +1,0 @@
-using Bunit;
-using Correlate;
-using FluentAssertions;
-using Fluxor;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Moq;
-using MudBlazor.Services;
-
-namespace Pizzeria.Store.Web.Tests.Unit;
-
-public class OrderPageTests : TestContext
-{
-    [Fact]
-    public void Order_WhenNoCurrentOrder_RedirectsToHome()
-    {
-        // Arrange
-        var mockApiClient = new Mock<IPizzeriaApiClient>();
-        var mockCorrelationContextAccessor = new Mock<ICorrelationContextAccessor>();
-        var mockLogger = new Mock<ILogger<OrdersEffects>>();
-
-        this.Services.AddSingleton(mockApiClient.Object);
-        this.Services.AddSingleton(mockCorrelationContextAccessor.Object);
-        this.Services.AddSingleton(mockLogger.Object);
-        this.Services.AddMudServices();
-        this.Services.AddFluxor(o => o.ScanAssemblies(typeof(OrdersState).Assembly));
-
-        // Act & Assert
-        // Since we don't have an active order, this should show a warning about no active order
-        var component = this.RenderComponent<Order>();
-        component.Markup.Should().Contain("No active order found");
-    }
-
-    [Fact]
-    public void Order_WhenPizzasLoading_ShowsSkeletonLoader()
-    {
-        // Arrange
-        var mockApiClient = new Mock<IPizzeriaApiClient>();
-        var mockCorrelationContextAccessor = new Mock<ICorrelationContextAccessor>();
-        var mockLogger = new Mock<ILogger<OrdersEffects>>();
-
-        this.Services.AddSingleton(mockApiClient.Object);
-        this.Services.AddSingleton(mockCorrelationContextAccessor.Object);
-        this.Services.AddSingleton(mockLogger.Object);
-        this.Services.AddMudServices();
-
-        // Set up Fluxor with initial state that has an active order
-        this.Services.AddFluxor(o =>
-        {
-            o.ScanAssemblies(typeof(OrdersState).Assembly);
-        });
-
-        // This test would be more effective with proper state setup
-        // For now, we'll test the basic structure
-        var component = this.RenderComponent<Order>();
-
-        // Assert - the component should render without throwing
-        component.Should().NotBeNull();
-    }
-
-    [Fact]
-    public void Order_WhenPizzasAvailable_ShowsPizzaCards()
-    {
-        // Arrange
-        var mockApiClient = new Mock<IPizzeriaApiClient>();
-        var mockCorrelationContextAccessor = new Mock<ICorrelationContextAccessor>();
-        var mockLogger = new Mock<ILogger<OrdersEffects>>();
-
-        var testPizzas = new[]
-        {
-            new PizzaDto(Guid.NewGuid(), "Margherita", "Classic tomato and mozzarella", 12.99m),
-            new PizzaDto(Guid.NewGuid(), "Pepperoni", "Pepperoni and cheese", 14.99m),
-        };
-
-        mockApiClient.Setup(x => x.GetPizzasAsync(It.IsAny<CancellationToken>())).ReturnsAsync(testPizzas);
-
-        this.Services.AddSingleton(mockApiClient.Object);
-        this.Services.AddSingleton(mockCorrelationContextAccessor.Object);
-        this.Services.AddSingleton(mockLogger.Object);
-        this.Services.AddMudServices();
-        this.Services.AddFluxor(o => o.ScanAssemblies(typeof(OrdersState).Assembly));
-
-        // Act
-        var component = this.RenderComponent<Order>();
-
-        // Assert - basic component structure
-
-        component.Markup.Should().Contain("Pizza Menu");
-    }
-}
