--- sample-tests/Pizzeria.Store.Web.Tests.Unit/TestContextExtensions.cs
@@ -1,29 +1,0 @@
-﻿using Bunit;
-using Correlate;
-using Fluxor;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Moq;
-using MudBlazor.Services;
-using Pizzeria.Store.Web.Infrastructure;
-using Pizzeria.Store.Web.Orders;
-
-namespace Pizzeria.Store.Web.Tests.Unit;
-
-internal static class TestContextExtensions
-{
-    public static void Setup(this TestContext testContext)
-    {
-        var mockApiClient = new Mock<IPizzeriaApiClient>();
-        var mockCorrelationContextAccessor = new Mock<ICorrelationContextAccessor>();
-        var mockLogger = new Mock<ILogger<OrdersEffects>>();
-
-        testContext.Services.AddSingleton(mockApiClient.Object);
-        testContext.Services.AddSingleton(mockCorrelationContextAccessor.Object);
-        testContext.Services.AddSingleton(mockLogger.Object);
-        testContext.Services.AddMudServices();
-        testContext.Services.AddFluxor(o => o.ScanAssemblies(typeof(OrdersState).Assembly));
-
-        testContext.JSInterop.Mode = JSRuntimeMode.Loose;
-    }
-}
